Use scope accounts/payments to track activity.
