--- src/store/modules/activity-selection.js
@@ -1,31 +1,0 @@
-import Vue from 'vue';
-import { SELECT_ACTIVITY } from '../mutation-types';
-
-const initialState = {
-  selectedActivity: null,
-};
-
-const getters = {
-  selectedActivity: state => () => { // eslint-disable-line
-    return state.selectedActivity;
-  },
-};
-
-const mutations = {
-  [SELECT_ACTIVITY](state, activity) {
-    Vue.set(state, 'selectedActivity', activity);
-  },
-};
-
-const actions = {
-  selectActivity({ commit }, activity) {
-    return commit(SELECT_ACTIVITY, activity);
-  },
-};
-
-export default {
-  state: initialState,
-  getters,
-  mutations,
-  actions,
-};
